FlaskApp/requirements.txt
alembic==1.9.4
click==8.0.4
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
python-dotenv==0.21.1
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.14.0